feat: rename withBroadcastState and add new hooks and validations
Added `time` to the update message to manage the `skipOlder` option.
Added `skipFirst` to avoid broadcast the initial load.
Added hooks `onStubImplementation`, `onSkipFirstBroadcast`, `onSkipDuplicatedBroadcast` and `onSkipOlder`.
The method `requestBroadcastState` is now public.
The method `_patchStateFromBroadcast` now receives as parameter the message.

BREAKING CHANGE: `withBroadcastSync` is renames to `withBroadcastState` and the default value to `requestState` is now `true`